requirements.txt
pytorchcv~=0.0.67
torch~=1.10.1
torchvision~=0.11.2
nltk~=3.6.7
scikit-learn~=0.24.2
networkx~=2.5.1
pytest~=7.0.1

opencv-python~=4.4.0.42
numpy~=1.19.5
Pillow~=8.4.0
nbdt~=0.0.4
tqdm~=4.19.9
Werkzeug~=2.0.3
setuptools~=59.6.0
matplotlib~=3.3.4
ttach~=0.0.3